LPS-74724 Create provider package and move it there
