+++ modules/apps/adaptive-media/adaptive-media-blogs-item-selector-web/src/main/java/com/liferay/adaptive/media/blogs/item/selector/web/internal/provider/AMBlogsItemSelectorViewReturnTypeProvider.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.blogs.item.selector.web.internal;
+package com.liferay.adaptive.media.blogs.item.selector.web.internal.provider;
 
 import com.liferay.adaptive.media.image.item.selector.AMImageFileEntryItemSelectorReturnType;
 import com.liferay.adaptive.media.image.item.selector.AMImageURLItemSelectorReturnType;
