+++ modules/apps/adaptive-media/adaptive-media-blogs-item-selector-web/src/test/java/com/liferay/adaptive/media/blogs/item/selector/web/internal/provider/AMBlogsItemSelectorViewReturnTypeProviderTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.document.library.item.selector.web.internal;
+package com.liferay.adaptive.media.blogs.item.selector.web.internal.provider;
 
 import com.liferay.adaptive.media.image.item.selector.AMImageFileEntryItemSelectorReturnType;
 import com.liferay.adaptive.media.image.item.selector.AMImageURLItemSelectorReturnType;
@@ -32,10 +32,38 @@ import org.powermock.api.mockito.PowerMockito;
 /**
  * @author Sergio González
  */
-public class AMDLItemSelectorViewReturnTypeProviderTest extends PowerMockito {
+public class AMBlogsItemSelectorViewReturnTypeProviderTest
+	extends PowerMockito {
 
 	@Test
-	public void testAddAMImageItemSelectorReturnTypesWithNonEmptyList()
+	public void testAddAMImageURLItemSelectorReturnTypeWithEmptyList()
+		throws Exception {
+
+		List<ItemSelectorReturnType> supportedItemSelectorReturnTypes =
+			new ArrayList<>();
+
+		ItemSelectorViewReturnTypeProvider
+			amBlogsItemSelectorViewReturnTypeProvider =
+				new AMBlogsItemSelectorViewReturnTypeProvider();
+
+		List<ItemSelectorReturnType> itemSelectorReturnTypes =
+			amBlogsItemSelectorViewReturnTypeProvider.
+				populateSupportedItemSelectorReturnTypes(
+					supportedItemSelectorReturnTypes);
+
+		Assert.assertEquals(
+			itemSelectorReturnTypes.toString(), 2,
+			itemSelectorReturnTypes.size());
+		Assert.assertTrue(
+			itemSelectorReturnTypes.get(0) instanceof
+				AMImageFileEntryItemSelectorReturnType);
+		Assert.assertTrue(
+			itemSelectorReturnTypes.get(1) instanceof
+				AMImageURLItemSelectorReturnType);
+	}
+
+	@Test
+	public void testAddAMImageURLItemSelectorReturnTypeWithNonemptyList()
 		throws Exception {
 
 		List<ItemSelectorReturnType> supportedItemSelectorReturnTypes =
@@ -46,18 +74,17 @@ public class AMDLItemSelectorViewReturnTypeProviderTest extends PowerMockito {
 		supportedItemSelectorReturnTypes.add(new URLItemSelectorReturnType());
 
 		ItemSelectorViewReturnTypeProvider
-			amDLItemSelectorViewReturnTypeProvider =
-				new AMDLItemSelectorViewReturnTypeProvider();
+			amBlogsItemSelectorViewReturnTypeProvider =
+				new AMBlogsItemSelectorViewReturnTypeProvider();
 
 		List<ItemSelectorReturnType> itemSelectorReturnTypes =
-			amDLItemSelectorViewReturnTypeProvider.
+			amBlogsItemSelectorViewReturnTypeProvider.
 				populateSupportedItemSelectorReturnTypes(
 					supportedItemSelectorReturnTypes);
 
 		Assert.assertEquals(
 			itemSelectorReturnTypes.toString(), 4,
 			itemSelectorReturnTypes.size());
-
 		Assert.assertTrue(
 			itemSelectorReturnTypes.get(0) instanceof
 				FileEntryItemSelectorReturnType);
@@ -69,35 +96,6 @@ public class AMDLItemSelectorViewReturnTypeProviderTest extends PowerMockito {
 				AMImageFileEntryItemSelectorReturnType);
 		Assert.assertTrue(
 			itemSelectorReturnTypes.get(3) instanceof
-				AMImageURLItemSelectorReturnType);
-	}
-
-	@Test
-	public void testAddAMImagetemSelectorReturnTypesWithEmptyList()
-		throws Exception {
-
-		List<ItemSelectorReturnType> supportedItemSelectorReturnTypes =
-			new ArrayList<>();
-
-		ItemSelectorViewReturnTypeProvider
-			amDLItemSelectorViewReturnTypeProvider =
-				new AMDLItemSelectorViewReturnTypeProvider();
-
-		List<ItemSelectorReturnType> itemSelectorReturnTypes =
-			amDLItemSelectorViewReturnTypeProvider.
-				populateSupportedItemSelectorReturnTypes(
-					supportedItemSelectorReturnTypes);
-
-		Assert.assertEquals(
-			itemSelectorReturnTypes.toString(), 2,
-			itemSelectorReturnTypes.size());
-
-		Assert.assertTrue(
-			itemSelectorReturnTypes.get(0) instanceof
-				AMImageFileEntryItemSelectorReturnType);
-
-		Assert.assertTrue(
-			itemSelectorReturnTypes.get(1) instanceof
 				AMImageURLItemSelectorReturnType);
 	}
