+++ modules/apps/adaptive-media/adaptive-media-item-selector-upload-web/src/test/java/com/liferay/adaptive/media/item/selector/upload/web/internal/provider/AMUploadViewReturnTypeProviderTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.item.selector.upload.web.internal;
+package com.liferay.adaptive.media.item.selector.upload.web.internal.provider;
 
 import com.liferay.adaptive.media.image.item.selector.AMImageFileEntryItemSelectorReturnType;
 import com.liferay.adaptive.media.image.item.selector.AMImageURLItemSelectorReturnType;
